+++ src/main/java/net/sf/oval/configuration/Configurer.java
@@ -17,7 +17,7 @@ public interface Configurer {
     * is invoked only once by the Validator, the very first time an object
     * of the given class needs to be validated. The constraint configuration
     * is then translated into an Validator internal format and cached.
-    * 
+    *
     * @param clazz the class to get the configuration for
     * @return The constraint configurations for the given class.
     * @throws InvalidConfigurationException in case of illegal configuration settings
@@ -27,7 +27,7 @@ public interface Configurer {
    /**
     * Returns the constraint configuration for the constraint set with the
     * given Id.
-    * 
+    *
     * @param constraintSetId the ID of the constraint set
     * @return The constraint configuration for the constraint set with the given Id.
     * @throws InvalidConfigurationException in case of illegal configuration settings