+++ src/main/java/net/sf/oval/guard/ParameterNameResolver.java
@@ -15,7 +15,7 @@ import net.sf.oval.exception.ReflectionException;
 public interface ParameterNameResolver {
    /**
     * Returns an empty String array if no parameters are declared.
-    * 
+    *
     * @param constructor the constructor
     * @return an array holding the parameter names of the given constructor.
     * @throws ReflectionException in case retrieving the parameter names fails
@@ -24,7 +24,7 @@ public interface ParameterNameResolver {
 
    /**
     * Returns an empty String array if no parameters are declared
-    * 
+    *
     * @param method the method
     * @return an array holding the parameter names of the given method.
     * @throws ReflectionException in case retrieving the parameter names fails